+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 ## Unreleased
 - Forbid `Value` and `Filterable` to be nested
 - Change Filter title format to allow upper case letters and underscore
+- Extends `IEnumerable` interface by `FiltersInterface` to implement both `Countable` and `IteratorAgregate` interfaces
+- Add `toArray` method to `FiltersInterface` to allow better debugging
 
 ## 1.0.0 - 2018-08-28
 - Initial version.